refactor: Refactor CLI to separate library
